+++ tests/Jroc.Test262.Tests/built-ins/InMemoryExecutionTestsBase.cs
@@ -25,7 +25,6 @@ public abstract class InMemoryExecutionTestsBase
             testName,
             _testCategory,
             name => GetJavaScriptAndSourcePath(name, sourceFilePath),
-            sourceFilePath,
             enableIRMetrics: true);
 
         Test262SharedAssertHarness.AssertNoOutput(testName, result.Output);